gltfpack: Refactor Basis support
Now that we only have a single mode, we can centralize the Basis
encoding support a little more; we no longer need abstract push/wait
functions etc.
