+++ gltf/gltfpack.h
@@ -293,9 +293,11 @@ void optimizeMaterials(cgltf_data* data, const char* input_path, std::vector<Ima
 bool readImage(const cgltf_image& image, const char* input_path, std::string& data, std::string& mime_type);
 bool hasAlpha(const std::string& data, const char* mime_type);
 bool getDimensions(const std::string& data, const char* mime_type, int& width, int& height);
+void adjustDimensions(int& width, int& height, const Settings& settings);
+const char* mimeExtension(const char* mime_type);
 
 #ifdef WITH_BASISU
-void encodeBasisInit(int jobs);
+void encodeInit(int jobs);
 bool encodeBasis(const std::string& data, const char* mime_type, std::string& result, const ImageInfo& info, const Settings& settings);
 void encodeImages(std::string* encoded, const cgltf_data* data, const std::vector<ImageInfo>& images, const char* input_path, const Settings& settings);
 #endif
